requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==22.1.0
autopep8==2.0.1
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
docopt==0.6.2
h11==0.14.0
idna==3.4
lxml==4.9.2
numpy==1.23.5
outcome==1.2.0
pandas==1.5.2
pipreqs==0.4.11
pycodestyle==2.10.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==13.10
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tornado==6.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
wsproto==1.2.0
yarg==0.1.9
